Renamed use_existing flag, made it more robust.
Video files are now generated in a temporary folder, and only moved to
the output folder when done. If use_cached_video_files is set, and the
target file already exists in the output folder, then it is not
regenerated.

Output file type is fored to mp4 for now, to allow the "exists()" check
to work. This could be made a parameter later if required.